requirements.txt
# --extra-index-url https://asia-south1-python.pkg.dev/glassy-totality-324307/covi-life-repository/simple
covi-life-core==0.0.2
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dataclasses==0.6
dnspython==2.2.1
email-validator==1.1.3
firebase-admin==5.2.0
Flask==2.0.3
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.0
Flask-request-params==0.3.0
Flask-Sockets==0.2.1
Flask-SocketIO==5.1.1
websocket-client==1.3.2
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-secret-manager==2.9.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.0.0
Mako==1.1.6
MarkupSafe==2.0.1
mongoengine==0.24.0
msgpack==1.0.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.0
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
reportlab==3.6.10
biosppy==0.8.0
matplotlib==3.5.2
flask-cors==3.0.10
pandas==1.4.3
xlrd==2.0.1
openpyxl==3.0.10
cryptography==37.0.4